requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
gunicorn==21.2.0
pytest==8.1.1
requests